+++ lib/main.dart
@@ -1,11 +1,9 @@
-import 'dart:async';
-
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:form_field_validator/form_field_validator.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+
 import 'app/get_it.dart';
+
+import 'pages/splashscreen.dart';
 
 void main() async {
   runApp(const MyApp());
@@ -33,2298 +31,4 @@ class MyCustomScrollBehavior extends MaterialScrollBehavior {
         PointerDeviceKind.stylus,
         PointerDeviceKind.unknown,
       };
-}
-
-class SplashScreen extends StatefulWidget {
-  const SplashScreen({super.key});
-
-  @override
-  State<SplashScreen> createState() => _SplashScreenState();
-}
-
-class _SplashScreenState extends State<SplashScreen> {
-  @override
-  void initState() {
-    super.initState();
-    Timer(
-        const Duration(seconds: 5),
-        () => Navigator.pushReplacement(context,
-            MaterialPageRoute(builder: (context) => const HomePage1())));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    setState(() {});
-    return Scaffold(
-      body: Center(
-        child: Container(
-          height: double.infinity,
-          width: double.infinity,
-          decoration: const BoxDecoration(
-            gradient: LinearGradient(
-              begin: Alignment.topLeft,
-              end: Alignment.bottomCenter,
-              stops: [0.1, 0.5, 0.7, 0.9],
-              colors: [
-                Color.fromARGB(255, 212, 199, 235),
-                Color.fromARGB(255, 151, 117, 250),
-                Color.fromARGB(255, 195, 172, 235),
-                Color.fromARGB(255, 151, 117, 250),
-              ],
-            ),
-          ),
-          child: Image.asset('images/logo.png'),
-        ),
-      ),
-    );
-  }
-}
-
-class HomePage1 extends StatefulWidget {
-  const HomePage1({super.key});
-
-  @override
-  State<HomePage1> createState() => _HomePage1State();
-}
-
-class _HomePage1State extends State<HomePage1> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Container(
-          height: double.infinity,
-          width: double.infinity,
-          decoration: const BoxDecoration(
-            gradient: LinearGradient(
-              begin: Alignment.topLeft,
-              end: Alignment.bottomCenter,
-              stops: [0.1, 0.5, 0.7, 0.9],
-              colors: [
-                Color.fromARGB(255, 212, 199, 235),
-                Color.fromARGB(255, 151, 117, 250),
-                Color.fromARGB(255, 195, 172, 235),
-                Color.fromARGB(255, 151, 117, 250),
-              ],
-            ),
-          ),
-          child: Stack(
-            children: [
-              Center(child: Image.asset('images/man.png')),
-              Padding(
-                padding: const EdgeInsets.all(20.0),
-                child: Align(
-                  alignment: Alignment.bottomCenter,
-                  child: Container(
-                    decoration: BoxDecoration(
-                        color: Colors.white,
-                        borderRadius: BorderRadius.circular(20)),
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        const Padding(
-                          padding: EdgeInsets.fromLTRB(42, 25, 42, 10),
-                          child: Text(
-                            'Look Good, Feel Good',
-                            style: TextStyle(
-                              fontSize: 30,
-                              fontWeight: FontWeight.w700,
-                            ),
-                          ),
-                        ),
-                        const Padding(
-                          padding: EdgeInsets.fromLTRB(25, 5, 25, 0),
-                          child: Text(
-                            'Create your individual & unique style and',
-                            style: TextStyle(
-                              fontSize: 18,
-                              fontWeight: FontWeight.w700,
-                              color: Color(0xFF8F959E),
-                            ),
-                          ),
-                        ),
-                        const Padding(
-                          padding: EdgeInsets.fromLTRB(25, 0, 25, 20),
-                          child: Text(
-                            'look amazing everyday.',
-                            style: TextStyle(
-                              fontSize: 18,
-                              fontWeight: FontWeight.w700,
-                              color: Color(0xFF8F959E),
-                            ),
-                          ),
-                        ),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            InkWell(
-                              onTap: () {
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                      builder: (context) => const Register()),
-                                );
-                              },
-                              child: Container(
-                                decoration: BoxDecoration(
-                                    color: const Color(0xFFF5F6FA),
-                                    borderRadius: BorderRadius.circular(10)),
-                                height: 60,
-                                width: 150,
-                                child: const Center(
-                                  child: Text('Man',
-                                      style: TextStyle(
-                                        fontSize: 18,
-                                        fontWeight: FontWeight.w500,
-                                        color: Color(0xFF8F959E),
-                                      )),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(10)),
-                            InkWell(
-                              onTap: () {
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                      builder: (context) => const Register()),
-                                );
-                              },
-                              child: Container(
-                                decoration: BoxDecoration(
-                                    color: const Color(0xFF9775FA),
-                                    borderRadius: BorderRadius.circular(10)),
-                                height: 60,
-                                width: 150,
-                                child: const Center(
-                                  child: Text(
-                                    'Women',
-                                    style: TextStyle(
-                                      fontSize: 18,
-                                      fontWeight: FontWeight.w500,
-                                      color: Color.fromARGB(255, 255, 255, 255),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                        const Padding(
-                          padding: EdgeInsets.all(15.0),
-                          child: Text(
-                            'Skip',
-                            style: TextStyle(
-                              fontSize: 18,
-                              fontWeight: FontWeight.w500,
-                              color: Color(0xFF8F959E),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class Register extends StatefulWidget {
-  const Register({Key? key}) : super(key: key);
-
-  @override
-  State<Register> createState() => _RegisterState();
-}
-
-enum Strength {
-  empty('Empty', 0 / 4, Colors.transparent),
-  weak('Weak', 1 / 4, Colors.red),
-  medium('Medium', 2 / 4, Colors.orange),
-  strong('Strong', 3 / 4, Colors.blue),
-  veryStrong('Very Strong', 4 / 4, Colors.green),
-  ;
-
-  final String text;
-  final double value;
-  final Color color;
-
-  const Strength(this.text, this.value, this.color);
-}
-
-bool _switchValue = true;
-
-class _RegisterState extends State<Register> {
-  bool isPressed = true;
-  Map userData = {};
-  final _formkey = GlobalKey<FormState>();
-
-  Strength _strength = Strength.empty;
-
-  void _updatePasswordStrength(String value) {
-    setState(() {
-      _strength = _calculatePasswordStrength(value);
-    });
-  }
-
-  Strength _calculatePasswordStrength(String value) {
-    if (value.isEmpty) {
-      return Strength.empty;
-    } else if (value.contains(RegExp(
-        r"^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$"))) {
-      return Strength.veryStrong;
-    } else if (value.contains(
-            RegExp(r"^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[a-zA-Z\d]{8,}$")) ||
-        value.contains(RegExp(
-            r"^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$"))) {
-      return Strength.strong;
-    } else if (value
-        .contains(RegExp(r"^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$"))) {
-      return Strength.medium;
-    } else {
-      return Strength.weak;
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        iconTheme: const IconThemeData(color: Color(0xFF1D1E20)),
-      ),
-      body: Form(
-        key: _formkey,
-        child: Column(
-          children: [
-            const Padding(
-              padding: EdgeInsets.fromLTRB(0, 50, 0, 100),
-              child: Text(
-                'Sign Up',
-                style: TextStyle(
-                  color: Colors.black,
-                  fontSize: 35,
-                  fontWeight: FontWeight.w700,
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: TextFormField(
-                validator: MultiValidator([
-                  RequiredValidator(errorText: 'Enter Username'),
-                  MinLengthValidator(3,
-                      errorText: 'Minimum 3 charecter filled Username'),
-                ]),
-                decoration: const InputDecoration(
-                  hintText: 'Enter Username',
-                  labelText: 'Username',
-                  prefixIcon: Icon(
-                    Icons.person,
-                    color: Color(0xFF9775FA),
-                  ),
-                  errorStyle: TextStyle(fontSize: 14.0),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: TextFormField(
-                onChanged: (value) => _updatePasswordStrength(value),
-                validator: MultiValidator([
-                  RequiredValidator(errorText: 'Enter Password'),
-                  PatternValidator(r"^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$",
-                      errorText:
-                          'Minimum eight characters, at least one letter and one number'),
-                ]),
-                obscureText: isPressed,
-                decoration: InputDecoration(
-                  suffix: Text(
-                    _strength.text,
-                    style: TextStyle(
-                      color: _strength.color,
-                      fontSize: 20,
-                    ),
-                  ),
-                  suffixIcon: IconButton(
-                    onPressed: () {
-                      setState(() {
-                        isPressed = !isPressed;
-                      });
-                    },
-                    icon: isPressed == true
-                        ? const Icon(Icons.visibility)
-                        : const Icon(Icons.visibility_off),
-                  ),
-                  hintText: 'Enter Password',
-                  labelText: 'Password',
-                  prefixIcon: const Icon(
-                    Icons.lock,
-                    color: Color(0xFF9775FA),
-                  ),
-                  errorStyle: const TextStyle(
-                    fontSize: 14.0,
-                  ),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: TextFormField(
-                validator: MultiValidator([
-                  RequiredValidator(errorText: 'Enter email address'),
-                  EmailValidator(errorText: 'Please correct email filled'),
-                ]),
-                decoration: const InputDecoration(
-                  hintText: 'Enter Email Address',
-                  labelText: 'Email Address',
-                  prefixIcon: Icon(
-                    Icons.email,
-                    color: Color(0xFF9775FA),
-                  ),
-                  errorStyle: TextStyle(fontSize: 14.0),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(15),
-              child: ListTile(
-                title: const Text(
-                  'Remember me',
-                  style: TextStyle(
-                      color: Color(0xFF1D1E20),
-                      fontSize: 18,
-                      fontWeight: FontWeight.w600),
-                ),
-                trailing: CupertinoSwitch(
-                  value: _switchValue,
-                  onChanged: (value) {
-                    setState(() {
-                      _switchValue = value;
-                    });
-                  },
-                ),
-              ),
-            ),
-            Expanded(
-              child: Align(
-                alignment: Alignment.bottomCenter,
-                child: Padding(
-                  padding: const EdgeInsets.fromLTRB(0, 0, 0, 0),
-                  child: InkWell(
-                    onTap: () {
-                      if (_formkey.currentState!.validate()) {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const Verification()),
-                        );
-                      }
-                    },
-                    child: Container(
-                      decoration: const BoxDecoration(color: Color(0xFF9775FA)),
-                      width: MediaQuery.of(context).size.width,
-                      height: 75,
-                      child: const Center(
-                        child: Text(
-                          'Sign Up',
-                          style: TextStyle(color: Colors.white, fontSize: 22),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class Verification extends StatefulWidget {
-  const Verification({Key? key}) : super(key: key);
-
-  @override
-  State<Verification> createState() => _VerificationState();
-}
-
-class _VerificationState extends State<Verification> {
-  Map userData = {};
-  final _formkey2 = GlobalKey<FormState>();
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        iconTheme: const IconThemeData(color: Color(0xFF1D1E20)),
-      ),
-      body: Form(
-        key: _formkey2,
-        child: Column(
-          children: [
-            const Padding(
-              padding: EdgeInsets.fromLTRB(0, 50, 0, 75),
-              child: Text(
-                'Verification Code',
-                style: TextStyle(
-                  color: Colors.black,
-                  fontSize: 35,
-                  fontWeight: FontWeight.w700,
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(10),
-              child: Image.asset('images/lock.png'),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: TextFormField(
-                style: const TextStyle(
-                    fontSize: 35,
-                    letterSpacing: 30,
-                    fontWeight: FontWeight.w700),
-                validator: MultiValidator([
-                  RequiredValidator(errorText: 'Enter Validator Cood'),
-                  PatternValidator(r"6521$", errorText: 'The code is wrong'),
-                ]),
-                decoration: InputDecoration(
-                  errorStyle: const TextStyle(
-                    fontSize: 25,
-                  ),
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                ),
-                textAlign: TextAlign.center,
-              ),
-            ),
-            const Padding(
-              padding: EdgeInsets.all(20),
-              child: Align(
-                alignment: Alignment.bottomCenter,
-                child: Text(
-                  'True Cood : 6521',
-                  style: TextStyle(
-                    color: Colors.black,
-                    fontSize: 20,
-                    fontWeight: FontWeight.w700,
-                  ),
-                ),
-              ),
-            ),
-            Expanded(
-              child: Align(
-                alignment: Alignment.bottomCenter,
-                child: Padding(
-                  padding: const EdgeInsets.fromLTRB(0, 0, 0, 0),
-                  child: InkWell(
-                    onTap: () {
-                      if (_formkey2.currentState!.validate()) {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const Screen9()),
-                        );
-                      }
-                    },
-                    child: Container(
-                      decoration: const BoxDecoration(color: Color(0xFF9775FA)),
-                      width: MediaQuery.of(context).size.width,
-                      height: 75,
-                      child: const Center(
-                        child: Text(
-                          'Confirm Code',
-                          style: TextStyle(color: Colors.white, fontSize: 22),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class Screen9 extends StatefulWidget {
-  const Screen9({super.key});
-
-  @override
-  State<Screen9> createState() => _Screen9State();
-}
-
-class _Screen9State extends State<Screen9> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color.fromARGB(255, 255, 255, 255),
-      bottomNavigationBar: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          const Padding(
-            padding: EdgeInsets.all(15.0),
-            child: Text(
-              'home',
-              style: TextStyle(
-                  color: Color(0xFF9775FA), fontWeight: FontWeight.w700),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(20.0),
-            child: Image.asset('images/h.png'),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(20.0),
-            child: Image.asset('images/bag.png'),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(20.0),
-            child: Image.asset('images/wallet.png'),
-          ),
-        ],
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
-            const Padding(padding: EdgeInsets.all(0)),
-            Padding(
-              padding: const EdgeInsets.all(15),
-              child: ListTile(
-                leading: IconButton(
-                  iconSize: 40,
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  icon: const Icon(
-                    Icons.arrow_circle_left_outlined,
-                    color: Color.fromARGB(255, 66, 62, 62),
-                  ),
-                ),
-                trailing: IconButton(
-                    iconSize: 40,
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => const Screen10()),
-                      );
-                    },
-                    icon: const Icon(Icons.shopping_bag_outlined,
-                        color: Color.fromARGB(255, 66, 62, 62))),
-              ),
-            ),
-            const Align(
-              alignment: Alignment.centerLeft,
-              child: Padding(
-                padding: EdgeInsets.fromLTRB(25, 10, 0, 10),
-                child: Text(
-                  'Hello',
-                  style: TextStyle(
-                      fontSize: 30,
-                      fontWeight: FontWeight.w600,
-                      color: Colors.black),
-                ),
-              ),
-            ),
-            const Align(
-              alignment: Alignment.centerLeft,
-              child: Padding(
-                padding: EdgeInsets.fromLTRB(27, 0, 0, 10),
-                child: Text(
-                  'Welcome to Laza.',
-                  style: TextStyle(
-                      fontSize: 15,
-                      fontWeight: FontWeight.w400,
-                      color: Color(0xFF8F959E)),
-                ),
-              ),
-            ),
-            Row(
-              children: [
-                Expanded(
-                  flex: 5,
-                  child: Padding(
-                    padding: const EdgeInsets.all(15.0),
-                    child: TextField(
-                      style: const TextStyle(fontSize: 20),
-                      decoration: InputDecoration(
-                        fillColor: const Color(0xFFF5F6FA),
-                        filled: true,
-                        hintText: 'Search',
-                        hintStyle: const TextStyle(fontSize: 25),
-                        prefixIcon: const Icon(
-                          Icons.search,
-                          size: 35,
-                          color: Color(0xFF8F959E),
-                        ),
-                        border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(20),
-                            borderSide: BorderSide.none),
-                      ),
-                    ),
-                  ),
-                ),
-                Expanded(
-                  flex: 1,
-                  child: Padding(
-                    padding: const EdgeInsets.fromLTRB(0, 15, 15, 15),
-                    child: Container(
-                      width: double.infinity,
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(20),
-                        color: const Color(0xFF9775FA),
-                      ),
-                      child: const Center(
-                        child: Padding(
-                          padding: EdgeInsets.all(17.0),
-                          child: Icon(
-                            Icons.mic_none,
-                            size: 30,
-                            color: Colors.white,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const Stack(
-              children: [
-                Align(
-                  alignment: Alignment.centerLeft,
-                  child: Padding(
-                    padding: EdgeInsets.fromLTRB(25, 10, 0, 10),
-                    child: Text(
-                      'Choose Brand',
-                      style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w600,
-                          color: Color(0xFF1D1E20)),
-                    ),
-                  ),
-                ),
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: Padding(
-                    padding: EdgeInsets.fromLTRB(25, 10, 25, 10),
-                    child: Text(
-                      'View All',
-                      style: TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.w600,
-                          color: Color(0xFF8F959E)),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            Padding(
-              padding: const EdgeInsets.all(10.0),
-              child: SingleChildScrollView(
-                scrollDirection: Axis.horizontal,
-                child: Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset('images/adidas.png'),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Adidas',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset("images/victor.png"),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Nike',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset("images/fila.png"),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Fila',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset("images/puma2.png"),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Puma',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset('images/adidas.png'),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Adidas',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset("images/victor.png"),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Nike',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset("images/fila.png"),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Fila',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    const Padding(padding: EdgeInsets.all(8)),
-                    Container(
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color(0xFFF5F6FA)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: Row(
-                          children: [
-                            Image.asset("images/puma2.png"),
-                            const Padding(padding: EdgeInsets.all(5)),
-                            const Text(
-                              'Puma',
-                              style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xFF1D1E20)),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            const Stack(
-              children: [
-                Align(
-                  alignment: Alignment.centerLeft,
-                  child: Padding(
-                    padding: EdgeInsets.fromLTRB(25, 10, 0, 10),
-                    child: Text(
-                      'New Arraival',
-                      style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w600,
-                          color: Color(0xFF1D1E20)),
-                    ),
-                  ),
-                ),
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: Padding(
-                    padding: EdgeInsets.fromLTRB(25, 10, 25, 10),
-                    child: Text(
-                      'View All',
-                      style: TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.w600,
-                          color: Color(0xFF8F959E)),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            Center(
-              child: Row(
-                children: [
-                  Expanded(
-                    flex: 5,
-                    child: Padding(
-                      padding: const EdgeInsets.all(20.0),
-                      child: InkWell(
-                        onTap: () {
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                                builder: (context) => const Screen10()),
-                          );
-                        },
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.start,
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                color: const Color.fromARGB(255, 242, 242, 242),
-                                borderRadius: BorderRadius.circular(15),
-                              ),
-                              child: Stack(
-                                alignment: Alignment.topLeft,
-                                children: [
-                                  Center(child: Image.asset('images/2.png')),
-                                  Align(
-                                    alignment: Alignment.topRight,
-                                    child: Padding(
-                                      padding: const EdgeInsets.all(10.0),
-                                      child: Image.asset('images/h.png'),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                            const Align(
-                              alignment: Alignment.bottomLeft,
-                              child: Padding(
-                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                                child: Text(
-                                  'Nike Sportswear Club',
-                                  maxLines: 2,
-                                  style: TextStyle(
-                                    fontSize: 15,
-                                    fontWeight: FontWeight.w400,
-                                    color: Color.fromARGB(255, 0, 0, 0),
-                                  ),
-                                ),
-                              ),
-                            ),
-                            const Align(
-                              alignment: Alignment.bottomLeft,
-                              child: Padding(
-                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                                child: Text(
-                                  'Fleece',
-                                  maxLines: 2,
-                                  style: TextStyle(
-                                    fontSize: 15,
-                                    fontWeight: FontWeight.w400,
-                                    color: Color.fromARGB(255, 0, 0, 0),
-                                  ),
-                                ),
-                              ),
-                            ),
-                            const Align(
-                              alignment: Alignment.bottomLeft,
-                              child: Padding(
-                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                                child: Text(
-                                  "\$ 99",
-                                  maxLines: 2,
-                                  style: TextStyle(
-                                    fontSize: 15,
-                                    fontWeight: FontWeight.w700,
-                                    color: Color.fromARGB(255, 0, 0, 0),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                  ),
-                  Expanded(
-                    flex: 5,
-                    child: Padding(
-                      padding: const EdgeInsets.all(20.0),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          Container(
-                            decoration: BoxDecoration(
-                              color: const Color.fromARGB(255, 242, 242, 242),
-                              borderRadius: BorderRadius.circular(15),
-                            ),
-                            child: Stack(
-                              alignment: Alignment.topLeft,
-                              children: [
-                                Center(child: Image.asset('images/4.png')),
-                                Align(
-                                  alignment: Alignment.topRight,
-                                  child: Padding(
-                                    padding: const EdgeInsets.all(10.0),
-                                    child: Image.asset('images/h.png'),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                'Trail Running Jacket Nike ',
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w400,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                'Windrunner',
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w400,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                "\$ 99",
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w700,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  )
-                ],
-              ),
-            ),
-            Center(
-              child: Row(
-                children: [
-                  Expanded(
-                    flex: 5,
-                    child: Padding(
-                      padding: const EdgeInsets.all(20.0),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          Container(
-                            decoration: BoxDecoration(
-                              color: const Color.fromARGB(255, 242, 242, 242),
-                              borderRadius: BorderRadius.circular(15),
-                            ),
-                            child: Stack(
-                              alignment: Alignment.topLeft,
-                              children: [
-                                Center(child: Image.asset('images/1.png')),
-                                Align(
-                                  alignment: Alignment.topRight,
-                                  child: Padding(
-                                    padding: const EdgeInsets.all(10.0),
-                                    child: Image.asset('images/h.png'),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                'Training Top',
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w400,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                'Nike Sport Clash',
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w400,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                "\$ 99",
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w700,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                  Expanded(
-                    flex: 5,
-                    child: Padding(
-                      padding: const EdgeInsets.all(20.0),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          Container(
-                            decoration: BoxDecoration(
-                              color: const Color.fromARGB(255, 242, 242, 242),
-                              borderRadius: BorderRadius.circular(15),
-                            ),
-                            child: Stack(
-                              alignment: Alignment.topLeft,
-                              children: [
-                                Center(child: Image.asset('images/3.png')),
-                                Align(
-                                  alignment: Alignment.topRight,
-                                  child: Padding(
-                                    padding: const EdgeInsets.all(10.0),
-                                    child: Image.asset('images/h.png'),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                'Trail Running Jacket Nike ',
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w400,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                'Windrunner',
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w400,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
-                              child: Text(
-                                "\$ 99",
-                                maxLines: 2,
-                                style: TextStyle(
-                                  fontSize: 15,
-                                  fontWeight: FontWeight.w700,
-                                  color: Color.fromARGB(255, 0, 0, 0),
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class Screen10 extends StatefulWidget {
-  const Screen10({super.key});
-
-  @override
-  State<Screen10> createState() => _Screen10State();
-}
-
-var x = 'images/6.png';
-
-class _Screen10State extends State<Screen10> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      body: Center(
-        child: SingleChildScrollView(
-          child: Column(
-            children: [
-              Center(
-                child: Container(
-                  width: double.infinity,
-                  color: const Color.fromARGB(255, 242, 242, 242),
-                  child: Stack(
-                    children: [
-                      Hero(
-                        tag: 'xxx',
-                        child: Center(
-                          child: Image.asset(x),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15),
-                        child: ListTile(
-                          leading: Container(
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(50),
-                                color: Colors.white),
-                            child: IconButton(
-                              iconSize: 30,
-                              onPressed: () {
-                                Navigator.pop(context);
-                              },
-                              icon: const Icon(
-                                Icons.arrow_circle_left_outlined,
-                                color: Color.fromARGB(255, 66, 62, 62),
-                              ),
-                            ),
-                          ),
-                          trailing: Container(
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(50),
-                                color: Colors.white),
-                            child: IconButton(
-                                iconSize: 30,
-                                onPressed: () {
-                                  Navigator.push(
-                                    context,
-                                    MaterialPageRoute(
-                                        builder: (context) => const Screen10()),
-                                  );
-                                },
-                                icon: const Icon(Icons.shopping_bag_outlined,
-                                    color: Color.fromARGB(255, 66, 62, 62))),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              const Stack(
-                children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 10, 0, 10),
-                      child: Text(
-                        "Men's Printed Pullover Hoodie",
-                        style: TextStyle(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w500,
-                            color: Color(0xFF8F959E)),
-                      ),
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.centerRight,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 10, 35, 10),
-                      child: Text(
-                        'Price',
-                        style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.w600,
-                            color: Color(0xFF8F959E)),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              const Stack(
-                children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 0, 0, 10),
-                      child: Text(
-                        "Nike Club Fleece",
-                        style: TextStyle(
-                            fontSize: 24,
-                            fontWeight: FontWeight.w700,
-                            color: Color(0xFF1D1E20)),
-                      ),
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.centerRight,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 0, 25, 10),
-                      child: Text(
-                        '\$ 120',
-                        style: TextStyle(
-                            fontSize: 24,
-                            fontWeight: FontWeight.w700,
-                            color: Color(0xFF1D1E20)),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              Padding(
-                padding: const EdgeInsets.all(10.0),
-                child: SingleChildScrollView(
-                  scrollDirection: Axis.horizontal,
-                  child: Row(
-                    children: <Widget>[
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/6.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/66.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/11-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/11.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/12-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/12.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/13-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/13.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/14-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/14.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/6.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/66.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/11-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/11.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/12-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/12.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/13-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/13.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(15.0),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              x = 'images/13-1.png';
-                            });
-                          },
-                          child: Row(
-                            children: [
-                              Image.asset('images/13.png'),
-                              const Padding(padding: EdgeInsets.all(5)),
-                            ],
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              const Stack(
-                children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 0, 0, 0),
-                      child: Text(
-                        "Size",
-                        style: TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.w700,
-                            color: Color(0xFF1D1E20)),
-                      ),
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.centerRight,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 0, 25, 0),
-                      child: Text(
-                        'Size Guide',
-                        style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.w400,
-                            color: Color(0xFF8F959E)),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              Padding(
-                padding: const EdgeInsets.all(10.0),
-                child: SingleChildScrollView(
-                  scrollDirection: Axis.horizontal,
-                  child: Row(
-                    children: <Widget>[
-                      Padding(
-                        padding: const EdgeInsets.all(5.0),
-                        child: Row(
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(10),
-                                color: const Color(0xFFF5F6FA),
-                              ),
-                              child: const Padding(
-                                padding: EdgeInsets.fromLTRB(30, 15, 30, 15),
-                                child: Text(
-                                  'S',
-                                  style: TextStyle(
-                                      fontSize: 25,
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w700),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(5)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(5.0),
-                        child: Row(
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(10),
-                                color: const Color(0xFFF5F6FA),
-                              ),
-                              child: const Padding(
-                                padding: EdgeInsets.fromLTRB(30, 15, 30, 15),
-                                child: Text(
-                                  'M',
-                                  style: TextStyle(
-                                      fontSize: 25,
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w700),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(5)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(5.0),
-                        child: Row(
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(10),
-                                color: const Color(0xFFF5F6FA),
-                              ),
-                              child: const Padding(
-                                padding: EdgeInsets.fromLTRB(30, 15, 30, 15),
-                                child: Text(
-                                  'L',
-                                  style: TextStyle(
-                                      fontSize: 25,
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w700),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(5)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(5.0),
-                        child: Row(
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(10),
-                                color: const Color(0xFFF5F6FA),
-                              ),
-                              child: const Padding(
-                                padding: EdgeInsets.fromLTRB(30, 15, 30, 15),
-                                child: Text(
-                                  'XL',
-                                  style: TextStyle(
-                                      fontSize: 25,
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w700),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(5)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(5.0),
-                        child: Row(
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(10),
-                                color: const Color(0xFFF5F6FA),
-                              ),
-                              child: const Padding(
-                                padding: EdgeInsets.fromLTRB(30, 15, 30, 15),
-                                child: Text(
-                                  '2XL',
-                                  style: TextStyle(
-                                      fontSize: 25,
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w700),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(5)),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.all(5.0),
-                        child: Row(
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(10),
-                                color: const Color(0xFFF5F6FA),
-                              ),
-                              child: const Padding(
-                                padding: EdgeInsets.fromLTRB(30, 15, 30, 15),
-                                child: Text(
-                                  '3XL',
-                                  style: TextStyle(
-                                      fontSize: 25,
-                                      color: Colors.black,
-                                      fontWeight: FontWeight.w700),
-                                ),
-                              ),
-                            ),
-                            const Padding(padding: EdgeInsets.all(5)),
-                          ],
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(25, 0, 0, 0),
-                  child: Text(
-                    "Description",
-                    style: TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.w700,
-                        color: Color(0xFF1D1E20)),
-                  ),
-                ),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(25, 0, 25, 0),
-                  child: Text(
-                    "The Nike Throwback Pullover Hoodie is made from premium French terry fabric that blends a performance feel with "
-                    "Read More..",
-                    style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w400,
-                        color: Color(0xFF8F959E)),
-                  ),
-                ),
-              ),
-              const Stack(
-                children: [
-                  Align(
-                    alignment: Alignment.centerLeft,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 25, 0, 10),
-                      child: Text(
-                        "Reviews",
-                        style: TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.w700,
-                            color: Color(0xFF1D1E20)),
-                      ),
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.centerRight,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 25, 25, 10),
-                      child: Text(
-                        'View All',
-                        style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.w400,
-                            color: Color(0xFF8F959E)),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              ListTile(
-                leading: CircleAvatar(
-                  child: Image.asset('images/15.png'),
-                ),
-                title: const Text(
-                  "Ronald Richards",
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                    color: Color(0xFF1D1E20),
-                  ),
-                ),
-                subtitle: const Row(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Icon(
-                      Icons.access_time_outlined,
-                      size: 18,
-                    ),
-                    Padding(padding: EdgeInsets.all(5)),
-                    Text('13 Sep, 2020')
-                  ],
-                ),
-                trailing: Image.asset('images/rate.png'),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(30, 0, 25, 20),
-                  child: Text(
-                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque malesuada eget vitae amet...",
-                    style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w400,
-                        color: Color(0xFF8F959E)),
-                  ),
-                ),
-              ),
-              const Stack(
-                children: [
-                  Align(
-                    alignment: Alignment.topLeft,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 0, 0, 25),
-                      child: Text(
-                        "Total Price",
-                        style: TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.w700,
-                            color: Color(0xFF1D1E20)),
-                      ),
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.centerRight,
-                    child: Padding(
-                      padding: EdgeInsets.fromLTRB(25, 0, 25, 25),
-                      child: Text(
-                        '\$125',
-                        style: TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.w700,
-                            color: Color(0xFF1D1E20)),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-      bottomNavigationBar: Padding(
-        padding: const EdgeInsets.fromLTRB(0, 0, 0, 0),
-        child: InkWell(
-          onTap: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => const SplashScreen()),
-            );
-          },
-          child: Container(
-            decoration: const BoxDecoration(color: Color(0xFF9775FA)),
-            width: MediaQuery.of(context).size.width,
-            height: 75,
-            child: const Center(
-              child: Text(
-                'Add to Cart',
-                style: TextStyle(color: Colors.white, fontSize: 22),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-TextEditingController dataname = TextEditingController();
-TextEditingController datades = TextEditingController();
-bool dess = false;
-
-class Reviews extends StatefulWidget {
-  const Reviews({super.key});
-
-  @override
-  State<Reviews> createState() => _ReviewsState();
-}
-
-class _ReviewsState extends State<Reviews> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        backgroundColor: Color(0xFFFEFEFE),
-        appBar: AppBar(
-          toolbarHeight: 100,
-          title: Center(
-              child: Text(
-            "Reviews",
-            style: TextStyle(color: Colors.black, fontWeight: FontWeight.w700),
-          )),
-          backgroundColor: Color(0xFFFEFEFE),
-        ),
-        body: Padding(
-          padding: const EdgeInsets.all(13.0),
-          child: Column(
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Image.asset('images/rate2.png', width: 150),
-                  Align(
-                    alignment: Alignment.centerRight,
-                    child: InkWell(
-                      onTap: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const AddReview()),
-                        );
-                      },
-                      child: Container(
-                        height: 50,
-                        width: 150,
-                        decoration: BoxDecoration(
-                          color: Color(0xFFFF7043),
-                          borderRadius: BorderRadius.circular(5),
-                        ),
-                        child: Padding(
-                          padding: EdgeInsets.fromLTRB(15, 0, 15, 0),
-                          child: Row(
-                            children: [
-                              Icon(
-                                Icons.edit_square,
-                                color: Colors.white,
-                              ),
-                              Text(
-                                'Add Review',
-                                style: TextStyle(
-                                    color: Colors.white, fontSize: 16),
-                              )
-                            ],
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              Visibility(
-                visible: dess,
-                child: ListTile(
-                  leading: CircleAvatar(child: Icon(Icons.account_box)),
-                  title: Text(
-                    getIt.get<SharedPreferences>().getString('pure_data') ??
-                        '  ',
-                    style: TextStyle(
-                      fontSize: 18,
-                      fontWeight: FontWeight.w500,
-                      color: Color(0xFF1D1E20),
-                    ),
-                  ),
-                  subtitle: const Row(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: [
-                      Icon(
-                        Icons.access_time_outlined,
-                        size: 18,
-                      ),
-                      Padding(padding: EdgeInsets.all(5)),
-                      Text('13 Sep, 2020')
-                    ],
-                  ),
-                  trailing: Image.asset('images/rate.png'),
-                ),
-              ),
-              Visibility(
-                visible: dess,
-                child: Align(
-                  alignment: Alignment.centerLeft,
-                  child: Padding(
-                    padding: EdgeInsets.fromLTRB(30, 0, 25, 20),
-                    child: Text(
-                      getIt.get<SharedPreferences>().getString('data_des') ??
-                          'hello',
-                      style: TextStyle(
-                          fontSize: 17,
-                          fontWeight: FontWeight.w400,
-                          color: Color(0xFF8F959E)),
-                    ),
-                  ),
-                ),
-              ),
-              ListTile(
-                leading: CircleAvatar(
-                  child: Image.asset('images/15.png'),
-                ),
-                title: const Text(
-                  "Jenny Wilson",
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                    color: Color(0xFF1D1E20),
-                  ),
-                ),
-                subtitle: const Row(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Icon(
-                      Icons.access_time_outlined,
-                      size: 18,
-                    ),
-                    Padding(padding: EdgeInsets.all(5)),
-                    Text('13 Sep, 2020')
-                  ],
-                ),
-                trailing: Image.asset('images/rate.png'),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(30, 0, 25, 20),
-                  child: Text(
-                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque malesuada eget vitae amet...",
-                    style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w400,
-                        color: Color(0xFF8F959E)),
-                  ),
-                ),
-              ),
-              ListTile(
-                leading: CircleAvatar(
-                  child: Image.asset('images/22.png'),
-                ),
-                title: const Text(
-                  "Ronald Richards",
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                    color: Color(0xFF1D1E20),
-                  ),
-                ),
-                subtitle: const Row(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Icon(
-                      Icons.access_time_outlined,
-                      size: 18,
-                    ),
-                    Padding(padding: EdgeInsets.all(5)),
-                    Text('13 Sep, 2020')
-                  ],
-                ),
-                trailing: Image.asset('images/rate.png'),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(30, 0, 25, 20),
-                  child: Text(
-                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque malesuada eget vitae amet...",
-                    style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w400,
-                        color: Color(0xFF8F959E)),
-                  ),
-                ),
-              ),
-              ListTile(
-                leading: CircleAvatar(
-                  child: Image.asset('images/21.png'),
-                ),
-                title: const Text(
-                  "Guy Hawkins",
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                    color: Color(0xFF1D1E20),
-                  ),
-                ),
-                subtitle: const Row(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Icon(
-                      Icons.access_time_outlined,
-                      size: 18,
-                    ),
-                    Padding(padding: EdgeInsets.all(5)),
-                    Text('13 Sep, 2020')
-                  ],
-                ),
-                trailing: Image.asset('images/rate.png'),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(30, 0, 25, 20),
-                  child: Text(
-                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque malesuada eget vitae amet...",
-                    style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w400,
-                        color: Color(0xFF8F959E)),
-                  ),
-                ),
-              ),
-              ListTile(
-                leading: CircleAvatar(
-                  child: Image.asset('images/20.png'),
-                ),
-                title: const Text(
-                  "Savannah Nguyen",
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                    color: Color(0xFF1D1E20),
-                  ),
-                ),
-                subtitle: const Row(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Icon(
-                      Icons.access_time_outlined,
-                      size: 18,
-                    ),
-                    Padding(padding: EdgeInsets.all(5)),
-                    Text('13 Sep, 2020')
-                  ],
-                ),
-                trailing: Image.asset('images/rate.png'),
-              ),
-              const Align(
-                alignment: Alignment.centerLeft,
-                child: Padding(
-                  padding: EdgeInsets.fromLTRB(30, 0, 25, 20),
-                  child: Text(
-                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque malesuada eget vitae amet...",
-                    style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w400,
-                        color: Color(0xFF8F959E)),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ));
-  }
-}
-
-class AddReview extends StatefulWidget {
-  const AddReview({super.key});
-
-  @override
-  State<AddReview> createState() => _AddReviewState();
-}
-
-class _AddReviewState extends State<AddReview> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color(0xFFFEFEFE),
-      appBar: AppBar(
-        toolbarHeight: 100,
-        title: Center(
-            child: Text(
-          "Add Review",
-          style: TextStyle(color: Colors.black, fontWeight: FontWeight.w700),
-        )),
-        backgroundColor: Color(0xFFFEFEFE),
-        iconTheme: CupertinoIconThemeData(
-          color: Colors.black,
-        ),
-      ),
-      body: Padding(
-        padding: EdgeInsets.fromLTRB(25, 25, 25, 0),
-        child: Column(
-          children: [
-            Align(
-              alignment: Alignment.topLeft,
-              child: Text(
-                "Name",
-                style: TextStyle(
-                    color: Colors.black,
-                    fontWeight: FontWeight.w700,
-                    fontSize: 18),
-              ),
-            ),
-            Padding(padding: EdgeInsets.all(5)),
-            TextField(
-              controller: dataname,
-              decoration: InputDecoration(
-                fillColor: Color(0xFFF5F6FA),
-                filled: true,
-                border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(10),
-                    borderSide: BorderSide.none),
-                hintText: 'Type your name',
-                errorStyle: TextStyle(fontSize: 14.0),
-              ),
-            ),
-            Align(
-              alignment: Alignment.topLeft,
-              child: Padding(
-                padding: EdgeInsets.only(top: 25),
-                child: Text(
-                  "How was your experience ?",
-                  style: TextStyle(
-                      color: Colors.black,
-                      fontWeight: FontWeight.w700,
-                      fontSize: 18),
-                ),
-              ),
-            ),
-            Padding(padding: EdgeInsets.all(5)),
-            TextField(
-              controller: datades,
-              minLines: 6,
-              keyboardType: TextInputType.multiline,
-              maxLines: 10,
-              decoration: InputDecoration(
-                fillColor: Color(0xFFF5F6FA),
-                filled: true,
-                border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(10),
-                    borderSide: BorderSide.none),
-                hintText: 'Describe your experience?',
-                errorStyle: TextStyle(fontSize: 14.0),
-              ),
-            ),
-          ],
-        ),
-      ),
-      bottomNavigationBar: Padding(
-        padding: const EdgeInsets.fromLTRB(0, 0, 0, 0),
-        child: InkWell(
-          onTap: () => showDialog<String>(
-            context: context,
-            builder: (BuildContext context) => AlertDialog(
-              title: Center(
-                  child: Text(
-                'Thank you for your review',
-                style: TextStyle(fontWeight: FontWeight.w700),
-              )),
-              content: Image.asset('images/thank-you.gif'),
-              actions: <Widget>[
-                Center(
-                  child: TextButton(
-                    onPressed: () {
-                      dess = true;
-                      setState(() {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const Reviews()),
-                        );
-                      });
-                      getIt
-                          .get<SharedPreferences>()
-                          .setString('pure_data', dataname.text);
-
-                      getIt
-                          .get<SharedPreferences>()
-                          .setString('data_des', datades.text);
-                    },
-                    child: const Text(
-                      'Next',
-                      style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w500,
-                          color: Color(0xFF9775FA)),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          child: Container(
-            decoration: const BoxDecoration(color: Color(0xFF9775FA)),
-            width: MediaQuery.of(context).size.width,
-            height: 75,
-            child: const Center(
-              child: Text(
-                'Submit Review',
-                style: TextStyle(color: Colors.white, fontSize: 22),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
 }
