+++ lib/pages/homepage1.dart
@@ -1,0 +1,157 @@
+import 'package:flutter/material.dart';
+
+import 'register.dart';
+
+class HomePage1 extends StatefulWidget {
+  const HomePage1({super.key});
+
+  @override
+  State<HomePage1> createState() => _HomePage1State();
+}
+
+class _HomePage1State extends State<HomePage1> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Container(
+          height: double.infinity,
+          width: double.infinity,
+          decoration: const BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomCenter,
+              stops: [0.1, 0.5, 0.7, 0.9],
+              colors: [
+                Color.fromARGB(255, 212, 199, 235),
+                Color.fromARGB(255, 151, 117, 250),
+                Color.fromARGB(255, 195, 172, 235),
+                Color.fromARGB(255, 151, 117, 250),
+              ],
+            ),
+          ),
+          child: Stack(
+            children: [
+              Center(child: Image.asset('images/man.png')),
+              Padding(
+                padding: const EdgeInsets.all(20.0),
+                child: Align(
+                  alignment: Alignment.bottomCenter,
+                  child: Container(
+                    decoration: BoxDecoration(
+                        color: Colors.white,
+                        borderRadius: BorderRadius.circular(20)),
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        const Padding(
+                          padding: EdgeInsets.fromLTRB(42, 25, 42, 10),
+                          child: Text(
+                            'Look Good, Feel Good',
+                            style: TextStyle(
+                              fontSize: 30,
+                              fontWeight: FontWeight.w700,
+                            ),
+                          ),
+                        ),
+                        const Padding(
+                          padding: EdgeInsets.fromLTRB(25, 5, 25, 0),
+                          child: Text(
+                            'Create your individual & unique style and',
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w700,
+                              color: Color(0xFF8F959E),
+                            ),
+                          ),
+                        ),
+                        const Padding(
+                          padding: EdgeInsets.fromLTRB(25, 0, 25, 20),
+                          child: Text(
+                            'look amazing everyday.',
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w700,
+                              color: Color(0xFF8F959E),
+                            ),
+                          ),
+                        ),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            InkWell(
+                              onTap: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => const Register()),
+                                );
+                              },
+                              child: Container(
+                                decoration: BoxDecoration(
+                                    color: const Color(0xFFF5F6FA),
+                                    borderRadius: BorderRadius.circular(10)),
+                                height: 60,
+                                width: 150,
+                                child: const Center(
+                                  child: Text('Man',
+                                      style: TextStyle(
+                                        fontSize: 18,
+                                        fontWeight: FontWeight.w500,
+                                        color: Color(0xFF8F959E),
+                                      )),
+                                ),
+                              ),
+                            ),
+                            const Padding(padding: EdgeInsets.all(10)),
+                            InkWell(
+                              onTap: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => const Register()),
+                                );
+                              },
+                              child: Container(
+                                decoration: BoxDecoration(
+                                    color: const Color(0xFF9775FA),
+                                    borderRadius: BorderRadius.circular(10)),
+                                height: 60,
+                                width: 150,
+                                child: const Center(
+                                  child: Text(
+                                    'Women',
+                                    style: TextStyle(
+                                      fontSize: 18,
+                                      fontWeight: FontWeight.w500,
+                                      color: Color.fromARGB(255, 255, 255, 255),
+                                    ),
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        const Padding(
+                          padding: EdgeInsets.all(15.0),
+                          child: Text(
+                            'Skip',
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w500,
+                              color: Color(0xFF8F959E),
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
